refactor: Move functions to new file
